+++ Chromatics/Models/SettingsModel.cs
@@ -60,6 +60,7 @@ namespace Chromatics.Models
         public List<YeelightAdoptedDevice> deviceYeelightAdoptedDevices { get; set; } = new();
         public bool deviceAlienwareEnabled { get; set; } = false;
         public List<AlienwareAdoptedDevice> deviceAlienwareAdoptedDevices { get; set; } = new();
+        public bool deviceDynamicLightingEnabled { get; set; } = false;
         public string deviceHueBridgeIP { get; set; } = "127.0.0.1";
         public string deviceHueBridgeClientKey { get; set; } = "";
         public double deviceHueBridgeBrightness { get; set; } = -1;